Add update method to StudentDAO and its implementation
The StudentDAO interface now includes a new method named 'update' and a corresponding implementation of this method has been added to StudentDAOImpl. The update method has been designed to take a 'Student' type object as an argument and merge it with the current state of the student in the database, effectively updating the student's information. The CrudDemoApplication class has also been modified to use this new update method.